currency/pyproject.toml
[tool.poetry]
name = "currency_sync"
version = "0.1.0"
description = ""
authors = ["freenux <freegly@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
mysql-connector-python = "^8.0.33"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
currency_sync = "currency_sync.main:main"
